Fix rand()
The seed should be unsigned long, otherwise you can
have overflow problems.

+++ lib/c/rand.c
@@ -2,7 +2,7 @@
 #undef rand
 #undef srand
 
-static int next;
+static unsigned long next;
 
 void
 srand(unsigned seed)
